Bugfix/Improvement: Improved packing algorithm for text glyphs.
Reduced the fragmentation of remaining texture space when a glyph is allocated space. This significantly reduces the number of nodes in the collection after inserting many glyphs. The previous algorithm fragmented the top right and bottom left quadrants of the texture so badly that they were never used and most allocations followed a diagonal line from top left to bottom right.