Bugfix: Loading a monochome font file or a grayscale font file in monochrome mode caused the GlyphPage.RenderGlyph function to fail.
I have renamed the function to RenderGrayscaleGlyph and created a new function called RenderMonoGlyph to handle mono fonts. There was a second bug where the width from GlyphMetrics was different from the width value from the FTBitmap data. Due to this and the fact we needed Pitch and PixelMode information from FTBitmap, I have changed the structure of GlyphInfo to hold the relevant information. Finally I also cleaned up the Font.TryGetGlyph function.